third_party/WebKit/LayoutTests/webaudio/ChannelMerger/audiochannelmerger-disconnect-expected.txt

Test ChannelMergerNode behavior on dynamic input change.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Channel #0 contains only the constant 1.
PASS Channel #1 contains all the expected values in the correct order: [1,0].
PASS The index of first zero in the channel #1 is equal to 11008.
PASS successfullyParsed is true

TEST COMPLETE
